Add Haptic back.  Fixes crash with help hint cycling.
Removing the defaults was a terrible idea.  This restores the Haptic binding so that Player can connect to the object, and not crash with null pointer exception.

Does not recreate binding itself, so nothing will happen upon haptic pulse.  There is no code-based API way to disable the haptics, which we don't want.