Get help hints shown at launch.
This was kind of tricky, and required a Coroutine to wait for the ControllerButtonHints to finally Initialize, before we can set the hints active.  Anything earlier is skipped because the hints are not collated yet.

Seems to work.  Works if the controllers are active at launch, and also if they are turned on after launch.  Possibly there will be some funny edge cases like left only being turned on, but at the moment we just need to hit the important pieces, like having them showing for the normal case of controllers both on at launch.