Enable ControllerHints script on LeftHand too.
If the controllers are off when you launch, then the right hand would show the hint, but on activation the left hand would be off.
This doubles the work, and has an extra coroutine running to look for init, but doesn't seem to cause any actual problems.

Tested in multiple scenarios including on/off controllers.  Seems to work correctly with no exceptions.