Add the Steam vdf files for SteamPipe uploader.
These are the definition files for the on-Jenkins SteamPipe uploader to use to send the bits to the proper depot.

There are two, Beta and Release.
Beta is expected to be changing a lot, every build will get pushed.
Release will only be pushed manually after validation.

They will be automatically pulled from the Katanga root folder by the Jenkins scripts, but keeping them here makes sure they are in version control.