Update Steam vdf files to switch depot to 1127315.
Make a total hash of the depots and builds, so as part of cleaning that up, I moved the Release/default branch to a new depot 1127315, from 312.

This changes the scripts to use the correct depot for uploads.
Still not sure these belong here, might make more sense to have them in Jobs on Jenkins.