Remove KeyboardListener from LaunchAndPlay.
This is a low-level technique, and I'm worried it will be flagged by antivirus as a keylogger.
Since we don't support full screen actions by keyboard right now, remove this and the mousewheel modification, to avoid confusion.

Commented out, not removed.  We can revisit if we decide to allow keyboard for screen movements.