Add Auto-Quit functionality to Katanga.
This creates a second thread to watch for the game exiting.  We can see it here by using the Deviare IsActive on the process.
We use a second thread because IsActive always waits a full second for each call.  This thus polls evey 1s to see if game has exited.

Upon exit of game, we call Application.Quit to exit Katanga.  Seems to work correctly, but a bit more complicated than desired.