Add new technique for hooking games.  Late hook, based on .exe name.
This will wait for 3DFM to launch the game instead of lauching it directly in Katanga.  That makes it possible to watch for the game.exe to show up, and when it is found, install hooks then, based on finding the Present call.  This works to bypass the launchers or other problems at launch time.

Breaks the DirectMode games like Tomb Raider.  Works to enable GTA5, and any Origin game.