Remove all KeyboardListener code.
As part of cleanup, move all these commented out chunks into the LowLevelKeyboardListener class.  That is no longer called, but is a good out of view storage spot in case we want to revisit this.

Also added back the required Copyright notice from DylanWeb, where the code came from originally.