Move RecenterHMD and HideFloor into ControllerActions.
These belong with the other ControllerActions, because they can be activated by controller actions.  There are also a couple of modifications to the environment at launch, changing floor to match playspace size, and RecenterHMD, so those are moved here as well.

Added Unity.Start method for the desired Recenter and floor sizing at Start.