Complete rewrite of LaunchAndPlay.cs for cleanup and new launch approach.
Since we are pushing full launch duty down to Katanga now, out of 3DFM, this needs some work, and now is a good time to restructure the whole thing.  It got to be a hella mess, and this is straightening it out.

All of the launching and deviare code is now in a new file, Game.cs, as a Game object that handles launching and communication with the game itself.

Mostly runs.  Nearly identical to previous behavior, so nothing introduced that I can see at present.  Needs a new launching mechanism for the normal dx11 path.