Continue refactor of Launch method.  Break DX9 out into it's own routine.
Moves the DX9 path to it's own sequence, so that we can also hook the Direct3DCreate9 and Direct3DCreate9Ex calls.  Not strictly necessary, as it should be handled by OnLoad init in the native deviare plugin, but might get us into some other games.

Started looking at the Kingdom of Amalur, which does not connect..  Reason is that it's a Steam double launch, and we are unloaded on first launch.  So not solved by this refactoring.