Fix command line argument parsing to properly clean up our stuff.
Bioshock Infinite was hanging at launch. It did not like having the katanga.exe path passed into it as the second parameter.
This change removes that dangling exe parameter, and also filters out all of the -- parameters we are passing.

This should leave solely the actual parameters that 3DFM has been sending.
The very first parameter will be the executable path for katanga, but should not matter, and is setup by the OS.  But only one, not two in a row.