Experimental code to enable 3Dmigoto software cursor.  Doesn't work.
Trying to get the Present hook to be at the right spot in the chain, right after 3Dmigoto.  But there doesn't seem to be a way to create a swapchain without 3Dmigoto hooking it and returning the wrong one.   At least for Frostpunk.  Worked for Civ6.

Even CreateDeviceAndSwapChain with only swapchain input comes back with wrong result, because 3Dmigoto has hooked the underlying calls.
No good way to do this here I think.