Experimental hook Present from Unity app instead, to get system Present address.
Not sure if this would work, didn't finish.  Address fetched is definitely from system32\d3d11.dll, and could be passed down to the game to use.

But, it's not clear if this is worth doing anyway, because the hook chain will be different by then, and could easily gank someone else's hook.

Best bet to solve cursor problem is to make a special 3DMigoto build I think.