First step to add saving screen location and state.
Using the PlayerPrefs storage for these settings.  This changes the rotation of the Player object at launch to match whatever the user saved on a Recenter command.

Handles never saved settings by doing a normal recenter at launch, but if they ever set a specific angle, we use that from then on.

Working, including rotating the floor back to normal.