Revert back to solely StartGameByeExeFile, to avoid DesktopGameTheater.
DGT does a force quit on Katanga so they can launch and take over the game launch.
But I found it only does this upon a normal Steam -appid launch. If we do the exe launch, they don't take over.

Moved all that down to Katanga so we could have better control, but looks like it was a waste of time.
At present, if we Steam launch and DGT is enabled, they force quit Katanga.  People are being confused, so best to revert to solely exe launching.