Change game.Launch into a coroutine, so it can launch async to environment.
Having the environment wedged while waiting for the .exe to be found was a bad experience, so this changes the behavior to setup the launch as a coroutine with a 100ms yield for each loop.  So it will now draw the environment in Update at full frame rate as expected, while we wait for the .exe procid to show up in the coroutine.  That is running on the same thread, right before Update, but at least does not block the main drawing.

Tested to work with all four main launch variants.

Removed the WaitForDX11 to make it wait for every variant.  For SpyMgr launches like DX9 and DirectMode, it will find it immediately, but causes no problems to keep same structure for all.