Add a GlobalScreen container object around Screen, so that we can RecenterHMD, but off center.
Finally! After days of experiments, finally hit upon the obvious solution, which was to make a container for the screen itself.  The container is centered around zero, with no offset, and can be rotated and translated independently from the Screen itself.  This allows the RecenterHMD operation to operate on the GlobalScreen container, moving it and rotating it to match the vrCamera.  While the contained Screen just goes along for the ride.

The sizing and moving operations now only operate on the local coordinates of the Screen itself, and is then independent of the GlobalScreen location, all done relative to that.

Both are saved and restored via PlayerPrefs.  If they hit any movement control, or recenter operation, it will save the state.

Works for every case now, and allows the screen to be offset from center so that playing at keyboard off to the side of the play area is now natural.