Restore the time delay after launch, before we start checking for exe.
For Battlefield3, we would catch the first bf3.exe launch, and it would exit, leaving us trying to inject into dead process.

Adding back an 8 second delay, before we look for the exe.  Gives it time to settle down and get stable. No longer looks like crap because this is now a coroutine, so the environment remains live.