Might solve the hang at ResizeBuffers problems.  Was using wrong pointer test.
The pointer test in Present was looking for the gGameTexture, but the semaphore we are using for the other side is gGameHandle.

We immediately set gGameHandle to null upon ResizeBuffers, so that the Unity side can know that it's gone and stop using it.  It does this check before drawing a frame.

Using gGameTexture was related, but could apparently be stale part way through the ResizeBuffer, and thus crash the drawing in Katanga.

Also pushed the CreateSharedTexture back to next Present, because as soon the gGameHandle goes null, we'll know to rebuild there too, and that keeps it in a single spot.

Initial testing suggests this might solve the problem.