Add MultiSampling directly into the sbsShader to help clean up shimmer.
This is recommended by Oculus developer support, and they give the multisampling routine.  This is an easy fix, but is probably expensive since it's done every pixel, every frame.

Seems to be worth it though.  Cleaned up shimmery white on black text, and partly cleaned up the shimmer in FFXIV.