Switch DX9 side to late binding for shared texture creation, to match DX11.
Doesn't make sense to be different for each side, so make it match the same approach.
Clean up the Reset operation to just clear gGameSharedHandle and release and null the shared surfaces.

This is in preparation for a mutex operation around creation and reset.
Still working- tested in BulletStorm and in connected.