Add full Mutex handling for game graphic setup during CreateDevice and Reset.
We have had hangs and crashes, which seem to be related to drawing while the graphics envinroment is torn down during resolution changes, or full screen changes.

This commit adds a specific named Mutex to block the drawing code from accessing possibly stale shared surfaces when resetting.  This is only for DX9 at the moment.

Seems to work correctly, getting blocking behavior as expected during resets, and otherwise is fully synchronized.  Works fine in Dishonored and Portal 2.