Update DX9 side to use clean Mutex to block VR side, when rebuilding graphics.
Long uphill road to work this out.  But it's now functional, and working for DX9 games.

Mutex is captured/released for every drawing during VR side, using the nativeplugin.
When the game side is building graphics, or doing a Reset to new graphics, we capture the mutex there, and thus block the VR side from being able to use the shared surface to draw.  It will revert to grey screen during this time.
Once setup sequence is complete, it releases mutex, allowing new drawing to happen on VR side.

Very hard to get this working right.  Cannot use late binding at Present call, because we need to patch the CreateDevice hook to make sure we build the right objects.  Cannot capture mutex there and release in Present, because those are on different threads for some games.