Add Mutex lock for DX11 side.  For both Create and ResizeBuffers.
Similar idea to DX9, but DX11 is complicated by us just injecting into the game, and not ever seeing original CreateDevice.  For that path, we setup the shared surface from Present, and use the mutex to lock out the VR side there.

For Reset on game resolution or full screen changes, we lock the mutex around the ResizeBuffers, and rebuild the shared surface immediately there, while the mutex is locked.  This leaves it with a valid gGameSharedHandle, which will then work for drawing in the VR side.  This might have a single bad frame of data, but there is not a good way to make this work better when we inject into the middle of running game.  Which itself is a must, because hooking the startup code has caused hangs.

Tested and working in Headlander.