Switch logging to fprintf to the Unity log file in LocalLow.
Prior logging to VS output is not working well, too complicated an environment, and not working for Mutex debugging.
This now puts the log into call sequence order, so we can see each part as it happens.
Will also likely be useful for end-user problems.

Also includes Unity change to not show stack crawl for simple log entries, only for errors and exceptions.