Switch all DX9 logging to 3Dmigoto LogInfo style logging, using _IONBF unbuffered logging.
This actually works!  So both the game side and the Katanga native C++ side can log to the same file, and mostly keep off each others toes.  Nothing is lost, and sequence is preserved, which is the key part for following mutex management.

Right now it's a hard coded path, and no DX11, but this proves it can work for this case, and be an ajunct logging that will be helpful going forward.

There does not appear to be any way to get the C# side to also properly log to that same file.  Too much abstraction, and the FileStream seems to lock out everything until it closes the file.  Not worth debugging, as stackoverflow says it cannot be done to mix C# logging with C++ logging.  And, we don't actually need that.  If we need or want C# logging we can just make a call into the native side and be done with it.