Fix Index Controller for hint text. Switch to triggers from grip.
Index controllers were not showing the hint text because the controller never showed up, even though they worked.  They didn't show up because the Pose was not bound to the controller.  This binds the pose to controller for both hands.

Also, grip was a poor choice for these controllers, too hard to use, to easy to trigger accidentally.  This change also modifies Recenter to be right hand trigger, and Hide Floor as left hand trigger.