Fix mutex bug of allowing Unity side in too early.
Sloppy coding there, and set the ownMutex true..  on the Release stage.  Not good.
Skip reporting the 'ok' bool on the ReleaseMutex.

This keeps the ownMutex strictly owned by CaptureMutex, and will block until it grabs it, and then follows through with fetching the shared handle.

Also change signature of ReleaseSetupMutex to ignore bool result.