Add PRISM.Sharpen asset to project.  Full screen sharpen, bought on Asset store.
Does a full screen Sharpen effect, with some options.  Based on suggestion from forum feedback that we really need a sharpen effect.

This is the simplest way to add this, drop in a Unity Asset bundle we bought on the Asset store.  Works in VR with single-pass Stereo, so seems like a good match.