Setup Sharpening button so hint and actions work.
Lots of tedious details here, including all three of the controllers that are listed. No WMR so it won't work there just yet.

Rename to ToggleSharpeningAction to be consistent with other settings.  Attach to the script, and also now log state changes as they are rare and interesting.

Set up left hand controller A or X button for the sharpening effect.  Left hand menu for Vive controller.

Works and makes a dramatic improvement, when sharpening is set high at 1.85.  Clearly a winner I think. Hope it's not very expensive on GPU.