Add first pass at keyboard and xbox controller UI handling.
This uses the Unity InputManager, and calls the same routine that the VR controllers do.
First pass here sets up the zoom in/out of the screen using PageUp/Dn and Xbox right stick up/down.

Requires that Katanga be in front, otherwise the events are not seen.  Seems right because we don't really want background actions interfering wtih the game.
Fully working here, needs to be extended to remaining UI.