Fix yet another multi-threading bug that kept the mutex from working properly over resolution change.
Did not mark the mutex as unowned, which meant that it could reset the graphics while in conflict with the game side.  Sigh.

Grab then Release, for real now.

Can conceivably get mutliple calls, so maybe it makes sense to do a loop on ReleaseSetupMutex until unowned.  Not sure about that.