Add a billboard with text on the left, for hints on Keyboard and Controller use.
Not sure this is a terrific idea, but adding a text billboard with help text on the left, so that they can see what all the keys and buttons are for changing the UI.

Seems to work OK, can be hidden easily, and all controls work in both keyboard and controller.
Separate hint info PlayerPref because they might never see it, or turn it off and never care, but want to see VR controller hints.