Fix failing Mutex lock one more time.
Seriously hard to get this right.  Unity is not making things any easier.

The looping was wrong for last commit on mutex, so that it would never free the mutex, and thus never get the game connection.  The game would error out, because the 1 second timeout for grabbing the mutex would fail, and it would exit.

This commit moves the GrabMutex to the top of Update, and kills the bad StartOfFrame coroutine, which was too late.  The ReleaseMutex still happens late, at WaitForEndOfFrame.

There is an additional change, because Unity does something stupid during startup and can call Update multiple times before calling to EndOfFrame, which is contrary to the documentation.  That would leave us with two Captures, one release, and thus lockout the game side.  To fix this, we've added an additional Release to the native C++ plugin, so that it will keep over-releasing until it catches up.  The second call should always fail, during normal running.

Back to connecting properly.  There is a non-zero chance this does what I've been trying to get it to do, which is lock out each side when rebuilding the graphics.