Experimental prototype code for hooking DX9Ex games.
All commented out, but was running, so worth a checkin before removing it.
Rather than make a duplicate sequence for DX9Ex, we'll switch that variant to the DX11 style where we wait after launch to inject into game to find Present.  Bypasses the daisy chain to fetch Present, because we don't need to hook and fix the other calls like CreateTexture, because they'll work correctly in DX9Ex.