Change startup to use the OnHookAdded call to setup the hooks for the game.
This allows us to choose what hooks to install, based on the game type passed in.  We'll know game type, because of the hook being added.  The hook itself is not used for anything except to startup the Agent and then install the proper API hooks.

This means we won't try to add a DX11 hook in a DX9 game, or vice versa.

Tested in Baynotta and Portal2 with success.