Add DX9Ex target for LaunchType, to specify delayed connection for DX9Ex games like Bayonetta.
For any game that we know uses DX9Ex, we can now connect directly using this target.  We can specify this in the 3DFM profile by having it be DX9+waitForExe specified.  WaitForExe makes no sense on DX9 games, because we have to patch Create9 before it's called.

For the DX9 launch case, we hook either Direct3DCreate9 or Direct3DCreate9Ex based on waitForExe.  That tells the Katanga plugin side that it needs to take the DX9Ex path.