Change launch to check for waitForExe, and skip launch if it exists.
To bypass launchers, we are having 3DFM launch the app using SteamAppID style.  Then, here in Katanga, we just waitForExe to show up before injecting.

This works for FFXIV. And should have no impact on all prior games.

Added DX9Ex launch to the launch types, and set it to StartGameByExeFile as well.  None presently use this.