Move waitForExe back to Katanga launch.  Forces SteamAppID launch.
This now looks for waitForExe to be set, and if so, it will force a Steam style launch, not exe.
Everything else that comes in will be forced to Exe style.

This allows us to specify a specific game to be SteamAppID launched, like FFXIV.  Works there.
Also works in GTA5, because we specify Exe launch in the profile, which is then unchanged.  So it launches the launcher, and still waitsForExe=gta5.exe

This seems like an OK compromise for now.  Bit of a threaded mess through the game types and launchers.

Tested Portal2, TombRaider, Prey, GTA5, FFXIV, OriDE, all are working, so that's all the primary game types.