Start of curved screen changes.  Using SBSShader to curve.
The approach seems to be to use the sbsShader and in the vertex shader, figure out how to do the curve we want.
This initial code is pulled in from a website, and has the basic pieces in place.  Definitely curves, but is upside down, and stretched badly.  Needs some work to smooth it out.  Still shows 3D, as moving the vertices doesn't affect or half and half drawing.

Right now the screen is still a Quad, and has no vertices to move, so this is commented out.