Connect support for OpenGL wrapped games!
This adds a new path through the launching, that looks for the DX9Ex connection that Helifax creates for his OpenGL wrapper.
He creates a 3D Vision DirectMode device, and sends the OpenGL data to it.  Connecting to that DirectMode device, and swapping eyes like we did with DX11 works here.

Adds a lot of complexity, but seems worthwhile.

Tested in SOMA, which was working correctly.  Requested on forums.