Add a ResetAllAction for both triggers held down at once, to clear PlayerPrefs.
This is a chord in the input manager, and both buttons must be down at the same time to do this relatively dangerous operation.
Calls DeleteAll on PlayerPrefs, then Start, to reset the environment back to defaults.
Also added same operation for the keyboard/controller section, using named inputs for Cycle and Recenter both down at once.

Seems to work, see it called in debugger, and environment resets.  No help text for this.