Revise waitForExe to not force launch type to Steam.
That was a bit of a hack to get FFXIV working, and now I've found there is a flag to force steam instead.

This commit simplifies the waitForExe use, and make it more general purpose.

Currently only used by FFXIV. Bayonetta would have used it, but has 3D entry problems.