Add better error logging to all plugins, including error code.
We are getting an NVAPI error for some users, and have not been logging the error code for FatalExit calls.
These are never supposed to happen, but something is wrong on their systems.
Adds the HRESULT to the FatalExit, logs that error, and then calls FatalExit now, so we'll be able to get the code.