Reorder launch for DX9 to include launch params, for Bioshock.
Bioshock fix requires -dx9 flag, and that was lost by previous code.
This commit changes the path so that we add launchArguments to the gamePath passed to Deviare launch, and it works.
Also skips the 8 second wait for anything other than the exe style launches.

This allows us to connect and play Bioshock.