Add a CreatePlane.cs script to generate a plane that suits our big screen needs.
http://wiki.unity3d.com/index.php?title=CreatePlane

This is a script from the UnityWiki, which just generates a plane using inputs specified.  A StereoScreen object is also added here by running the script, which is 32 triangles wide by 18 triangles tall, double the 16x9 screen we want as default.  That gives us more vertices to work with, and a smoother curve.  Some 500 vertices is not going to break the GPU bank, even when we are setting them in the sbsShader.

Could make this a generate on the fly, and then a static screen instead of modiyfing the vertices every frame in the sbs shader, but it's not likely to make any performance difference, and it's simpler this way.