Change sbsShader to make screen curved.
Stricly speaking this does not need to be done live in the shader, we can generate a static curved screen, but this is fairly simple and the 500 or so vertices are not going to be measurably expensive for a given frame.

Code is setup to generate a circular curve, as opposed to exponential or any other curve, because human eyesight apparently follows a circular curve as well.  Not positive this is right, but the google searches don't show anything better.

Needs work, but this is a good start.