Replace screen with improved version that is 1x1 in size.
Because of the way the old code is setup, we need to be able to scale by 8x4.5 to get a normal size screen.  That will be the defaults used in the field.

This replaces the screen with a 32x18 triangle screen, but set to 1x1 size so we can scale properly.  After much thrashing around on the -4.5.  We have that setup and working too.

Tested with Headlander, and it's showing on curved screen, looks right.  Curve is live and moves as slider is moved.  Controller actions still working.