Fix curve settings for screen.  Fully working now.
This adds some logic to convert a linear 0..10 curve setting into something that will curve the screen in a linear way.  The math was farily complex to work out.

Works to map the 0..10 to a infinite..4 radius setting, and seems pretty good. 4 is the max, because otherwise the sqrt goes negative and clips.

Importantly, this folows the same 8x-4.5 setup as the old screen, so that none of the controller bindings and logic need to change for moving and resizing.  That was quite tricky to get working, but means that all the saved states in the field will continue to work.