Setup ControllerActions to handle curving actions.
Add the CurvingScreen coroutine, to change it dynamically during controller actions.  Move the circle calculations into this routine, so that it calculates new vertices off the original mesh at each call.  The saved value is the curve coefficient 0..10, but probably stops working at 5 because it will clip.

Not quite right, but functional here, and the screen can be rotated with Recenter and still work for curving.

Original vertices are saved at Start so we can make a fresh clone as needed for a new curve coefficient.