Fix UpdateCurve to only use local scaling.
Mistake there was that TransformPoint does in fact include the container object, and thus any rotation that was set via Recenter would make screen skew instead of curve.

Just changing this to only do the scaling manually, not using any functions.  We use localScale, which is just the transform for StereoScreen, nothing else.  It will never be rotated or translated in X, so this should work in all scenarios.