Fix distortion in the curved screen.
This fixes the distortion by recalculating every x,z vertices on the mesh.  The calculation trick was to use the angles specified, not the starting x,z locations.  This distributes the x's evenly across the arc of the screen, with a fixed angle of 11.25 degrees between each segment, so each segment is exactly the same length as before.

With this change, we now get perfectly smooth images.  Still distorted because of the curve itself, but static images look good.

Currently hard coded to a 180 degree screen.