Add --wait-time parameter handling for launching.
We can now specify the wait time to be used for a given game in their profile.  This handles that parameter, and sets the wait time loop before it finds the launched exe and injects.
Opens up some testing possibilities for end users too.

Helps in NierAutomata and also Kingdom Come.