Update help text for all controllers.
Took awhile, but figured out how to determine which controller is active, and change the text so it makes more sense.
The Index and WMR controller has extra dpad, so it's cleaner to push curve screen off to that.  On OG Vive and Rift, it's just joystick, so it will be up/down/left/right.

Tested all but Rift. Looks good except for Index, which jams the text together.  Not worth the time to fix that.