Add a barrel distortion to the Mesh after curving.
The idea here is to counter the distortion introduced by the curve of the screen, which is pincushion distortion, by making a countering barrel distortion.

The size of the distortion constant is not clear, but this commit is the basic code for getting this to work.  Very hard to find the right idea for code to make this happen.
Right now it is applying distortion after the vertices have moved to their curve positions, which works because the distortion is a delta from their curved location.

Seems to work, the screen is curved, but does not actually look curved in VR.