Remove barrel distortion experimental code.
Wrong branch, but we'll just remove the code here.

This was an interesting experiment, to add barrel distortion to counteract the pincushion distortion introduced by having the screen curve.  This example code is running and works for the default 0.5 curve.

It works to remove distortion but also to remove the curve.  So visually it looks flat screen.  Screen is curved as seen in Unity editor, but visually it's flat.  That seems like the wrong approach, because the end-users apparently care that it LOOKS curved, and don't care about distoration.  So in the interest of letting the customer do stupid stuff that makes no sense, we'll leave it as is, and remove this barrel fix.