Add a "Stereo Pictures" subfolder to UnityScreenApp, for set of demo pics.
Small set to start with here, but gives a location for the storage.  Change from Nvidia.img folder, to this set, as being more obvious, and also that people doing demo probably don't have 3D Vision or screenshot folder.

Add a copy command in the Release build, to copy folder to Release output.