Switch build back to katanga.exe from HelixVisionDemo.exe as output.
Some dependency that I cannot find on the name of the exe breaks the build if it comes from steam.  Works fine on all builds, except for steam download.

Don't really care that it has proper exe name, and this is a waste of time.  Just switching back to katanga.exe.