Fix last bug in Pause function, to start properly.
All bindings are set as Toggle, so we need to only use the 'active' flag here, not toggle again.
And, the Toggle is for Pause On, so we invert that for the 'playing' flag.

This fixes the pause/resume to properly pause and resume on single clicks of right hand grip.