Remove all marginal screenshots from Demo.  Add lots of BonusShots.
Getting to a minimal set that is all good quality, good separation and convergence.  Needs to be improved, but OK for first ship.

Add another 40 shots or so to the BonusShots folder.  Bit of a waste of space, 60M total for shots now.  But might be interesting for people to play with.  None of these shots are good enough for the main slideshow, either a weakly framed shot, or too much depth.