Update default prefs for screen size and location, to center at head height.
Screen was too high to avoid clipping against floor.
This is a tweak of defaults to make screen small enough so that it won't clip, but then also bring it closer so that the field of view does not change.
Params are roughly half the size they were, but visually very similar.

Improves the code for returning defaults to have them all in single routines, in one place, rather than multiple copies prone to getting out of sync.