Comment out Deviare injection for DX11 path.
Experimental, but seems to be working.  Will still need the DX9 injection path.

Needs more work to smooth out the DX11 launches, maybe set those back to Steam style directly out of 3DFM, because the IPC will now be picked up by Katanga directly whether it comes from 3Dmigoto or the GamePlugin.

Improves the startup handling, but checking for nulls during launch.  Since the game via 3Dmigoto is completely asynchronous, we can't be sure when the IPC file mapping will arrive, nor the mapped view.  So, it now handles those being null if Katanga is already drawing and looking for game buffers.  This is all using a late-binding approach directly out of Present, so the timing will be unknown.