Allow any DX11 game to launch without any process checks.
For DirectConnect, we don't need any Deviare injection, so this removes the finding of process ID, and waiting for process to launch.

We are going to be using the model of just waiting for the FileMappedIPC to appear, as the notification that it is ready to run.

Restores launching by SteamAppID instead of exe.  Not sure this is best, because this has the side effect of always bringing up the SteamVR dashboard now.