Move all DX11 cases into a single switch case, and unify all three DX11 types.
All three DX11 are launched out of 3DFM now, so this will just wait for the named exe to arrive in the process list, and save the process id to watch for exit.

All three DX11 types: Steam, EpicID, DX11Exe are handled the same now. And will connect via the 3Dmigoto named file IPC.