Move screen down and shrink it to fit for seated position.
This puts center of screen at y=1m, which should be a reasonable seated height.

Shrinks the screen to match that lowered position, so that it does not clip through the ground. x=3.556, y=-2, still 16:9.

Then moves the screen closer to make it bigger in the field of view,  From 3m to 2.5m

So now screen is 2.5m away at 3.556 meters by 2 meters in size.  4.08 meters diagonal, 160 inches diagonal.