Add the uDesktopDuplication package to  Katanga.
This is from https://github.com/hecomi/uDesktopDuplication as a free project that implements the Win10 desktop duplication API in Unity.  Using 1.6 version as 1.7 is not compatible with Unity 2017.4

We can add this project, connect the texture to the main window, and get desktop views inside VR.  Worth a bit of work to get this functional.