Add the Texture.cs to the StereoScreen object.
This is now attached directly to the screen, as it matches the code of uDesktopDuplication better.  It will automatically fetch the MeshFilter and the sbsMaterial and display the desktop data directly into the default parameters.

We don't need their shader to do this, and in fact their shader is more primitive than ours, so we'll leave our sbsMaterial as it is.

Some small tweaks to the Texture.cs to disable fetching of parameters we don't care about, Bend, Forward, Cull, Radius, Width, Thickness. We don't want their code to modify our screen, and these otherwise give errors, because those parameters don't exist in our sbsShader.