Decide what mode to run during LaunchAndPlay.Start.
Three modes, slideshow, normal vr, and desktop duplication.  Decides based on input params which mode to set active.

To make desktopduplication work, the script starts off disabled, then is enabled in Start if there are no input params.  And also the LaunchAndPlay is disabled, so that it doesn't interfere with uDesktopDuplication.

Tested as working in all three modes.  Although in DesktopDuplication we are also in stereo because the shader is looking for stereo.