Add new Unlit/shader2D shader for uDesktopDuplication use.
We want to use our shader to avoid conflicts with theirs, but don't want to pay any performance hit to check if 2D or not 2D.  So this shader will be used whenever we switch modes at LaunchAndPlay.

It's a normal 2D shader, but includes the Oculus MSAA for shimmer fixing. Works to show the desktop on the big screen when DesktopDuplication is active.